drission-page/pyproject.toml
[project]
name = "drission-scraper"
version = "0.1.0"
description = "DrissionPage signup automation worker"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "drissionpage==4.1.1.2",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.uv]
package = false
